requirements.txt
python-fasthtml
uvicorn>=0.29
openai==1.52.0
transformers
torch